Fix acceptance test: use proper ENV variables
In roder to make the provider works with new nested blocks per product
we need to make sure to configure the block in the tests in order to
activate the SDK to pick up ENV variables. In this case block can be
empty but it is required to make the entire system work.