Refactor: move requests handling into the clients module
Moving all the requests into the `clients` we make sure to have only
single place where we have to take care of the clients related
issues/upgrades etc.
All the requests jsut wrapped into ApiClient, which handles
Marshaling/Unmarshaling of the data and returns proper types to the
caller to process them.