Identify missing/deleted resources
Additionaly check response http status, and if it's 404 we know that
resource is missing and we must set ID to empty string ("") in the state
so the TF will be able to re-create the resources again.

TF will keep the state and resources up-to-date, even if those were
accidentally removed.