Add downsample to ILMs
There is a parameter for the hot, warm and cold phases called
downsample.

According to the description:

"Roll up documents within a fixed interval to a single summary document.
Reduces the index footprint by storing time series data at reduced
granularity."

As of 0.11.0 there is no support for this parameter in the proivider,
this PR adds it.