Fix fleet_enrollment_tokens returning empty token set in some case
The data source is not handling pagination at all. This fix the case
when kibana server got more than 20 fleet enrollment tokens and the
provider is not paginating the result when searching for a match.

Fetching all tokens case (empty policyID in input) is still not handled.
Just trying to push a fix with minimum effort for now.

Partially fixes #528.