remove space_id paramter from private locations (#733)
fix for #726 

private location is agnostic to kibana spaces. The field space_id was previously incorrectly used, because its value was a `namespace` of agent policy (data stream).

This PR removes the field. As there is no way to update a private location via API (only replace), there is no benefit to introduce `namespace` field for private locations.