Don't populate settings during read
There's a fair bit of complexity around handling computed settings. I'll look at adding this back in a follow up PR, but want to get this resource in a working state so we can make a new release.

This covers the case where settings are applied as part of an index template