Apply Kibana config defaults respecting mutually exclusive auth values
Instead of using the defaults as a starting point we apply them after building the config object to any fields that are not set. In doing so we respect fields that are mutually exclusive like api_key and username. Applying defaults at the end makes it easier to check we aren't accidentally building an invalid config.